Some updates in the prompts and default models.
Also added gpt-4 as model for some of the characters and
added a system message that tells the number of tokens
used with each message.
I used to find myself wondering how many tokens I have spend in the conversation and how many more I could use, this solves that although I would like to make the conversation manager summarize the conversation so it can continue.